script/upload_resource/requirements.txt
flake8
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
mysql-connector-python
